ui/companion: move the core picker filter and the core-information rows out of the Qt companion
Phase 2, ninth subsystem: everything the Qt companion still computed
from core_info itself.

- LoadCoreWindow::initCoreList filtered the core list by gathering the
  content's extension plus every archive member's extension into a
  QStringList and matching it against each core's supported_extensions
  string. That was a Qt re-implementation of core_info's own matcher;
  companion_core_installed_cores_supporting(content_path) now calls
  core_info_list_get_supported_cores() - the exact rule the menu uses,
  directories included - and reports how many cores lead the list. The
  table is built from the installed-core accessors (a _version accessor
  is added; _name falls back to the core file's base name, as the table
  did) and hides the rows past that count. MainWindow::loadContent's
  CORE_SELECTION_ASK path no longer walks archives; it passes the path.

- The core-information collector (~240 lines: name / version / system /
  authors / permissions / licences / extensions / firmware status with
  per-row colouring / notes) was already toolkit-free C living in
  ui_qt.cpp under a qt_ prefix. It moves verbatim into the core as
  companion_core_core_info_rows() with companion_core_info_row_status;
  MainWindow::getCoreInfo keeps only the HTML/CSS decoration of the
  firmware rows.

Both are what the native companions' Load Core dialog and core-info
panel will consume.

Verified: full Qt 6 retroarch links and starts headless with the qt
companion; companion_core.c and the consumers of companion_core.h pass
tools/mingw_syntax_check.sh (mingw -Wall -Wextra clean).
